fix(cmd): serve and dev should work
